Move pulling an import to 'calkit sync import'
'calkit update' is the namespace for editing a declaration: every other
subcommand updates a spec, a config, or metadata, and 'calkit update
dataset' already carries --imported-from-url, --imported-from-doi and
--imported-from-git-url for editing exactly this record. Sitting beside
those, 'update import' read as "change the import" when it fetches bytes.

'calkit sync' is described as syncing with external systems, which is
what pulling an import is, so it moves there alongside git, dvc and
overleaf. Behavior is unchanged: 'calkit sync import <path>' or --all,
with --force for a file edited since it was fetched.
