+++ docs/calkit-yaml.md
@@ -641,11 +641,11 @@ Data from a Git repo that isn't a Calkit project.
 
 #### `_GitSource`
 
-| Parameter  | Type        | Required | Default | Description                                                                                                                                                                                                                                                                                                                                           |
-| ---------- | ----------- | -------- | ------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `repo_url` | str         | yes      |         | Clone URL of the repo the data came from.                                                                                                                                                                                                                                                                                                             |
-| `rev`      | str \| None | no       | null    | Deprecated; the commit an import resolved to is recorded in .calkit/imports.json, which is committed alongside it. This file says what to follow, which a person writes; that one says where following it led, which the tool works out. Still read for entries written before the split, and moved across the next time 'calkit update import' runs. |
-| `path`     | str \| None | no       | null    | Path within that repo, if it isn't the whole thing.                                                                                                                                                                                                                                                                                                   |
-| `ref`      | str \| None | no       | null    | Branch, tag, or commit to follow when refreshing this, e.g., 'main'. Optional: an entry that names none is refreshed from the repo's default branch. 'rev' still records the commit actually fetched, so the entry says both what it tracks and what it got.                                                                                          |
+| Parameter  | Type        | Required | Default | Description                                                                                                                                                                                                                                                                                                                                         |
+| ---------- | ----------- | -------- | ------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `repo_url` | str         | yes      |         | Clone URL of the repo the data came from.                                                                                                                                                                                                                                                                                                           |
+| `rev`      | str \| None | no       | null    | Deprecated; the commit an import resolved to is recorded in .calkit/imports.json, which is committed alongside it. This file says what to follow, which a person writes; that one says where following it led, which the tool works out. Still read for entries written before the split, and moved across the next time 'calkit sync import' runs. |
+| `path`     | str \| None | no       | null    | Path within that repo, if it isn't the whole thing.                                                                                                                                                                                                                                                                                                 |
+| `ref`      | str \| None | no       | null    | Branch, tag, or commit to follow when refreshing this, e.g., 'main'. Optional: an entry that names none is refreshed from the repo's default branch. 'rev' still records the commit actually fetched, so the entry says both what it tracks and what it got.                                                                                        |
 
 <!-- AUTO-GENERATED: CALKIT-YAML-KEYS:END -->
